Feature parity for the Zoom queue
This push corresponds to the update that brought feature parity to the zoom queue. Currently the files are set up for the in lab help queue, There are some changes that need to be made if they are going to be put in the zoom help queue. These changes are live on the websites. The biggest change aside from feature parity is that both queues are now visible on the TA end. That means we can easily keep an eye on both queues.